config/jina-reranker/requirements.txt
# =============================================================================
# Jina Reranker Service Dependencies
# =============================================================================

# Web Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0

# HTTP Client
httpx==0.28.1

# Configuration
pydantic==2.11.9
pydantic-settings==2.0.0
pyyaml==6.0.2

# Caching
redis==5.0.0

# Monitoring
prometheus-client==0.23.1

# Logging
structlog==24.4.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-multipart==0.0.12